Implement RemoveComponent + Return raw pointers
GetComponent() and AddComponent() returns a raw ptr instead of a shared_ptr. This will prevent longer lifetime than expected if another object is holder the returned shared_ptr.

I think that a weak_ptr is the best returning type that i can pick, because the raw pointers are not linked with the shared_ptr anymore (no status update if all the shared_ptr got deleted). This will return in errors if the user intend to use a component while he is deleted. But during other classes I was told to avoid weak_ptr, it's a sign of bad design.

This is a perfect question for during next class.