Applied personal feedback
GameObjects can be killed, the gameObjects and the components get
destroyed inside the late update, not in the update. This way seems more
logical personal to me. Constructor of Component class is also changed,
the owner was previously a pointer, changed this to a reference. A
reference can't be a nullptr so we can avoid a nullcheck during
component creation.